refactor droppable interface
