+++ packages/controls/src/widget_string.ts
@@ -13,14 +13,62 @@ import {
     uuid
 } from './utils';
 
-import * as _ from 'underscore';
+import {
+    DOMWidgetView
+} from '@jupyter-widgets/base';
 
+import * as _ from 'underscore';
 
 /**
  * Class name for a combobox with an invlid value.
  */
 const INVALID_VALUE_CLASS = 'jpwidgets-invalidComboValue';
 
+/**
+* https://www.typescriptlang.org/docs/handbook/mixins.html
+**/
+function applyMixins(derivedCtor: any, baseCtors: any[]) {
+    baseCtors.forEach(baseCtor => {
+        Object.getOwnPropertyNames(baseCtor.prototype).forEach(name => {
+            Object.defineProperty(derivedCtor.prototype, name, Object.getOwnPropertyDescriptor(baseCtor.prototype, name));
+        });
+    });
+}
+
+/**
+ * Handler for widget drop events
+ * Can be used to create droppable widgets
+ */
+class Droppable extends DOMWidgetView {
+    _handle_drop(event) {
+        event.preventDefault();
+        // var data = Array.from(event.dataTransfer.items, item => item.getAsString())
+        var data = [];
+
+        for (var i=0; i < event.dataTransfer.types.length; i++) {
+          var t = event.dataTransfer.types[i];
+          data.push({type: t, value: event.dataTransfer.getData(t)})
+        }
+
+        var datamap = new Object();
+
+        for (var i=0; i < event.dataTransfer.types.length; i++) {
+          var t = event.dataTransfer.types[i];
+          datamap[t] = event.dataTransfer.getData(t);
+        }
+
+        console.log(event.dataTransfer);
+        this.send({event: 'drop', data: datamap});
+    }
+
+    /**
+     * Dictionary of events and handlers
+     */
+    events(): {[e:string] : string;}
+    {
+        return {'drop': '_handle_drop'};
+    }
+}
 
 export
 class StringModel extends CoreDescriptionModel {
@@ -125,7 +173,7 @@ class LabelModel extends StringModel {
 }
 
 export
-class LabelView extends DescriptionView {
+class LabelView extends DescriptionView implements Droppable {
     /**
      * Called when view is rendered.
      */
@@ -164,39 +212,13 @@ class LabelView extends DescriptionView {
         return super.update();
     }
 
-    /**
-     * Dictionary of events and handlers
-     */
-    events() {
-        // TODO: return typing not needed in Typescript later than 1.8.x
-        // See http://stackoverflow.com/questions/22077023/why-cant-i-indirectly-return-an-object-literal-to-satisfy-an-index-signature-re and https://github.com/Microsoft/TypeScript/pull/7029
-        return {'drop': '_handle_drop'};
-    }
+    _handle_drop: () => void
+    events: () => {[e:string] : string;}
 
-    /**
-     * Handles when the button is clicked.
-     */
-    _handle_drop(event) {
-        event.preventDefault();
-        // var data = Array.from(event.dataTransfer.items, item => item.getAsString())
-        var data = [];
 
-        for (var i=0; i < event.dataTransfer.types.length; i++) {
-          var t = event.dataTransfer.types[i];
-          data.push({type: t, value: event.dataTransfer.getData(t)})
-        }
-
-        var datamap = new Object();
-
-        for (var i=0; i < event.dataTransfer.types.length; i++) {
-          var t = event.dataTransfer.types[i];
-          datamap[t] = event.dataTransfer.getData(t);
-        }
-
-        console.log(event.dataTransfer);
-        this.send({event: 'drop', data: datamap});
-    }
 }
+
+applyMixins(LabelView, [Droppable])
 
 export
 class TextareaModel extends StringModel {
